provide data from a fake data-source
